requirements.txt
kivy==2.1.0
numpy==1.23.5
matplotlib==3.6.2
pandas==1.5.2
scipy==1.9.3
pyusb==1.2.1